requirements.txt
Django
djangorestframework
celery
djoser
django-filter
pillow
django-extra-fields
python-dotenv
psycopg2-binary
pep8
pytest
pylint
pycodestyle
flake8
gunicorn


-r requirements/requirements-optionals.txt
-r requirements/requirements-testing.txt
-r requirements/requirements-documentation.txt
-r requirements/requirements-packaging.txt